refactored utilites to separate modules re #16
