+++ ebagis/utils/zipfile.py
@@ -1,4 +1,24 @@
-
+def zip_directory(directory_path, zip_path):
+    """Takes an input directory path and an output zip path
+    and zips the contents of the directory into a zipfile at
+    output path location. The name of the output zipfile
+    should be included in the zip_path argument."""
+    import os
+    import zipfile
+    with zipfile.ZipFile(zip_path, 'w', zipfile.ZIP_DEFLATED) as zipf:
+        for root, dirs, files in os.walk(directory_path):
+            for f in files:
+                rel_dir = os.path.relpath(root, directory_path)
+                # first arg is file to zip (root/f), and the
+                # second arg is the archive name (rel_dir/f)
+                # (second arg is path realtive to the root of the AOI
+                # to keep whole higher level directory structure from
+                # getting zipped)
+                zipf.write(os.path.join(root, f), os.path.join(rel_dir, f))
+            # NEED TO USE THIS FOR EMPTY DIRS
+            #for d in dirs:
+              #  zips.writestr(zipfile.ZipInfo('empty/'), '')
+    return zip_path
 
 
 def unzipfile(zipf, unzipdir):
